undo some adding
